+++ ui/panels-plugin/package.json
@@ -29,8 +29,6 @@
     "lint:fix": "eslint --fix src --ext .ts,.tsx"
   },
   "dependencies": {
-    "@atlaskit/pragmatic-drag-and-drop": "^1.4.0",
-    "@atlaskit/pragmatic-drag-and-drop-hitbox": "^1.0.3",
     "@mui/x-data-grid": "^7.6.2",
     "@perses-dev/components": "0.48.0",
     "@perses-dev/core": "0.48.0",
